modularized musicians modal and button, made GigCard component into server component
